Breaking change: Refactor API and Introduce Blob Cache
Summary:

* Refactored the API for improved robustness.
* Introduced a composite LRU+TTL cache mechanism.

Details:

The API has been revamped to enhance its robustness. A notable change is the
implementation of the Blob protocol for returned blobs, which now allows
accessing the underlying data as a byte array, a file, or an input stream.

To achieve flexibility and extensibility, specific backend implementations
(e.g., postgres, sql server, s3) are injected through dependency inversion,
conforming to the Backend protocol.

Blobs are now cached locally using a combination of LRU and TTL caches. The LRU
cache ensures that frequently used blobs are cached, while the TTL cache aids
in maintaining cache coherencei (i.e.: eventual consistency) in distributed
systems. It's not perfect, though; the cache should be avoided when immediate
consistency is required, and the API provides the means for that.

Users have the option to provide custom callbacks to monitor cache behavior.
These functions will be invoked on each cache hit, miss, or eviction event,
allowing users to compute the hit rate using the formula hit / (hit + miss).

Additionally, comprehensive unit tests have been added to verify the
functionality of the system.